Adding relRowNum to RowPart class
As it turned out it is very inconvenient to have only absolute row
number in row parts. Now there are two values: relative row index in
paragraph and absolute in the document. This requires rethink of
behaviour of RowPartsBuilder class. In addition Paragraph class get
roParts field with all row parts which it consists of.